add extends met

+++ src/main/java/com/DS/demo/repositories/TableRepo.java
@@ -3,5 +3,5 @@ package com.DS.demo.repositories;
 import com.DS.demo.models.TableEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TableRepo extends JpaRepository<TableEntity,Integer> {
+public interface TableRepo extends JpaRepository<TableEntity,Long> {
 }
