+++ src/main/java/com/DS/demo/repositories/TicketRepo.java
@@ -3,5 +3,5 @@ package com.DS.demo.repositories;
 import com.DS.demo.models.TicketEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketRepo extends JpaRepository<TicketEntity,Integer> {
+public interface TicketRepo extends JpaRepository<TicketEntity,Long> {
 }
